+++ src/pic/simulator/parser/commands/Incf.java
@@ -32,7 +32,7 @@ public class Incf extends Command
 
 	@Override
 	public void execute(PicProcessor proc) {
-		byte val = proc.getMemoryControl().getAt(arg0);
+		short val = proc.getMemoryControl().getAt(arg0);
 		val++;
 
 		if(arg1==0)
