+++ src/pic/simulator/specialfunctionregisters/Portb.java
@@ -11,7 +11,7 @@ import pic.simulator.pins.Pin;
 public class Portb extends SpecialFunctionRegister
 {
     private final HashMap<Integer, IOPin> pins;
-    private byte                          value;
+    private short                          value;
 
     public Portb(PicProcessor processor)
     {
@@ -31,9 +31,9 @@ public class Portb extends SpecialFunctionRegister
     }
 
     @Override
-    public void setValue(byte value)
+    public void setValue(short value)
     {
-        this.value = (byte) (value & 0x1f);
+        this.value = (short) (value & 0x1f);
 
         if ((value & 0b00000001) > 0)
             pins.get(Pin.RB0).setInternally();
@@ -77,7 +77,7 @@ public class Portb extends SpecialFunctionRegister
     }
 
     @Override
-    public byte getValue()
+    public short getValue()
     {
         value = 0;
